requirements.txt
boto3
openai
feedparser
python-dotenv